+++ source/extensions/isaacsim.sensors.rtx.nodes/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.1.3] - 2026-06-14
+### Added
+- `register_scalar_colored_point_cloud_writer` utility to color RTX point clouds by a scalar field (distance, intensity, ...).
+
 ## [0.1.2] - 2026-06-11
 ### Fixed